python-engine/requirements.txt
fastapi
uvicorn
google-genai
akshare
duckdb
pandas
pandas-ta
pydantic
python-dotenv
tenacity
loguru
